Minor fix
git-svn-id: svn+ssh://circa.ecs.vuw.ac.nz/u/staff/djp/projects/repository/tuttepoly/trunk@4722 cab769db-1af3-0310-8bf3-f2d376eea66d

+++ tutte/tutte.cpp
@@ -1292,8 +1292,8 @@ void write_graph_sizes(ostream &out) {
   out << "CACHE GRAPH SIZE DATA" << endl;
   out << "=====================" << endl;
   out << "V\t#Graphs (%)\t#MultiGraphs (%)" << endl;
-  vector<int> counts;
-  vector<int> mcounts;
+  vector<long> counts;
+  vector<long> mcounts;
   int nmgraphs=0;
   int ngraphs=0;
   // first, count the lengths
